llama: fix FFI type-contract violations at the libffi boundary
Six defects over 17 call sites where the cif, the C prototype and the Go
variable disagree. No cgo, so none of them fail at build time.

Return buffers:

- ggml_backend_cpu_buffer_type was bound TypeVoid but C returns an 8-byte
  pointer. libffi writes nothing for a void return, so it returned 0 on
  every call; the NULL reaches GGML_ASSERT(buft) and abort()s model load.
  Measured 0 -> a valid non-NULL buffer type.
- ModelRopeFreqScaleTrain, VocabGetScore read float returns via ffi.Arg.
  libffi writes only the low 4 bytes, and float32(uint64) converts rather
  than reinterprets. Measured 1.229783e+19 -> 1.0.
- LoadModeFromStr used a 4-byte LoadMode return buffer; libffi always
  stores a full 8-byte ffi_arg, so it wrote 4 bytes past the variable.

Argument widths:

- size_t min_keep passed as uint32 in top-p, min-p, typical and XTC.
  libffi read 8 bytes from a 4-byte local, splicing adjacent memory into
  the high word; min_keep >= 2^32 makes llama.cpp silently disable the
  filter.
- size_t buf_size passed as int32 at 7 sites (model.go x4, lora.go x3),
  so the snprintf bound was billions instead of 128/32768 and never
  smaller. Overflow primitive against the Go heap. SplitPath and
  SplitPrefix were already correct.
- MemorySeqDiv passed an 8-byte Go int into a 4-byte TypeSint32 slot.

No exported signatures change; the samplers keep uint32 and widen
internally.

Tests: add TestGGMLBackendCpuBufferType. RopeFreqScaleTrain only logged
its value and VocabGetScore only rejected negatives, which a positive
1.2e19 passes; both now assert magnitude.

Verified on llama.cpp b10219: pkg/... green, audit goes from 264/265,
502/514, 272/276 to 265/265, 514/514, 276/276 with 0 skips.

+++ pkg/llama/lora.go
@@ -128,7 +128,7 @@ func AdapterMetaValStr(adapter AdapterLora, key string) (string, bool) {
 	}
 	buf := make([]byte, 32768)
 	b := unsafe.SliceData(buf)
-	bLen := int32(len(buf))
+	bLen := uint64(len(buf))
 
 	keyPtr, _ := utils.BytePtrFromString(key)
 	var result ffi.Arg
@@ -167,7 +167,7 @@ func AdapterMetaKeyByIndex(adapter AdapterLora, i int32) (string, bool) {
 	}
 	buf := make([]byte, 128)
 	b := unsafe.SliceData(buf)
-	bLen := int32(len(buf))
+	bLen := uint64(len(buf))
 
 	var result ffi.Arg
 	adapterMetaKeyByIndexFunc.Call(
@@ -194,7 +194,7 @@ func AdapterMetaValStrByIndex(adapter AdapterLora, i int32) (string, bool) {
 	}
 	buf := make([]byte, 32768)
 	b := unsafe.SliceData(buf)
-	bLen := int32(len(buf))
+	bLen := uint64(len(buf))
 
 	var result ffi.Arg
 	adapterMetaValStrByIndexFunc.Call(
